requirements.txt
# SpoonOS Core Framework
# spoon-core>=0.1.0

# LangGraph for Graph Agent
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# LLM Providers
openai>=1.0.0
anthropic>=0.34.0
google-generativeai>=0.3.0

# HTTP and API
requests>=2.31.0
httpx>=0.27.0

# Web3 (Optional)
web3>=6.0.0
eth-account>=0.13.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Environment Management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Logging
loguru>=0.7.0

# Memory (Optional)
mem0ai>=0.1.0

# Tools
tavily-python>=0.3.0

# Async Support
asyncio>=3.4.3
aiohttp>=3.9.0

# Utilities
pydantic>=2.0.0
rich>=13.0.0
